requirements.txt
python-telegram-bot==13.15
gspread
oauth2client
PyMuPDF
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2